Create a custom hook to fetch games
